Static

+++ DummyEngine/src/Dummy/ImGui/Layer/ImGuiLayer.cpp
@@ -76,7 +76,7 @@ namespace Dummy
         ImGuiIO& io = ImGui::GetIO();
         Application& app = Application::Get();
 
-        io.DisplaySize = ImVec2(app.GetWindow().GetWidth(), app.GetWindow().GetHeight());
+        io.DisplaySize = ImVec2(static_cast<float>(app.GetWindow().GetWidth()), static_cast<float>(app.GetWindow().GetHeight()));
 
         ImGui::Render();
         ImGui_ImplOpenGL3_RenderDrawData(ImGui::GetDrawData());
