+++ DummyEngine/src/Platform/Windows/Input/WindowsInput.cpp
@@ -37,7 +37,7 @@ namespace Dummy
 
         glfwGetCursorPos(window, &xPos, &yPos);
 
-        return {xPos, yPos};
+        return { static_cast<float>(xPos), static_cast<float>(yPos)};
     }
 
     float WindowsInput::GetMouseXImpl()
